requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==41.0.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-allauth==0.56.1
django-autoslug==1.9.9
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-formtools==2.4.1
django-htmx==1.14.0
django-import-export==3.2.0
django-jsoneditor==0.2.4
django-render-block==0.9.2
et-xmlfile==1.1.0
Faker==18.4.0
fonttools==4.41.1
html5lib==1.1
idna==3.4
install==1.3.5
lxml==4.9.3
Markdown==3.4.4
MarkupPy==1.14
mysqlclient==2.2.0
numpy==1.25.1
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pycparser==2.21
pydyf==0.7.0
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pypdf==3.13.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.4
webencodings==0.5.1
websocket-client==1.5.1
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.2
